backend/requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
pycparser==2.23
PyMySQL==1.1.2
sqlparse==0.5.5
tzdata==2025.3
Pillow==12.1.0
python-dotenv==1.0.1
